code adjustments, lint fixing
