+++ packages/account-lookup-svc/package.json
@@ -34,7 +34,7 @@
     "publish": "npm run clean && npm run build && npm version patch && npm publish --tag latest --access public",
     "lint": "eslint ./src",
     "lint:fix": "eslint ./src --fix",
-    "test:unit": "jest --passWithNoTests --testMatch **/test/unit/*.test.ts" ,
+    "test:unit": "jest --passWithNoTests --detectOpenHandles --testMatch **/test/unit/*.test.ts" ,
     "test:integration": "jest --passWithNoTests --testMatch **/test/integration/**/*.test.ts",
     "test:coverage": "jest --passWithNoTests --coverage --coverageThreshold='{}' --testMatch **/test/unit/**/*.test.ts",
     "test:coverage-check": "jest --coverage --testMatch **/test/unit/**/*.test.ts"
@@ -47,7 +47,7 @@
     "@mojaloop/auditing-bc-public-types-lib": "^0.0.7",
     "@mojaloop/logging-bc-client-lib": "^0.0.10",
     "@mojaloop/logging-bc-public-types-lib": "^0.0.3",
-    "@mojaloop/platform-shared-lib-nodejs-kafka-client-lib": "^0.0.8"
+    "@mojaloop/platform-shared-lib-nodejs-kafka-client-lib": "^0.1.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
