+++ packages/domain/package.json
@@ -29,7 +29,6 @@
     "audit:check": "npm npm audit",
     "dep:check": "ncu -e 2",
     "dep:update": "ncu -u",
-    "publish": "npm run clean && npm run build && npm version patch && npm publish --tag latest --access public",
     "lint": "eslint ./src",
     "lint:fix": "eslint ./src --fix",
     "test:unit": "jest --passWithNoTests --testMatch **/test/unit/**/*.test.ts ",
