+++ packages/domain/src/aggregate.ts
@@ -48,304 +48,302 @@ import { IOracleFinder, IOracleProvider, IParticipantService} from "./interfaces
 import { AccountLookUpEventsType, IAccountLookUpMessage, IParticipant, ParticipantAssociationRequestReceived, ParticipantDisassociationRequestReceived, ParticipantQueryReceived, PartyInfoAvailable, PartyQueryReceived } from "./types";
 export class AccountLookupAggregate  {
 	private readonly _logger: ILogger;
-    private readonly _oracleFinder: IOracleFinder;
+	private readonly _oracleFinder: IOracleFinder;
 	private readonly _oracleProviders: IOracleProvider[];
-    private readonly _messageProducer: IMessageProducer;
-    private readonly _participantService: IParticipantService;
-    
+	private readonly _messageProducer: IMessageProducer;
+	private readonly _participantService: IParticipantService;
+	
 
 	constructor(
 		logger: ILogger,
-        oracleFinder:IOracleFinder,
-        oracleProviders:IOracleProvider[],
-        messageProducer:IMessageProducer,
-        participantService: IParticipantService
+		oracleFinder:IOracleFinder,
+		oracleProviders:IOracleProvider[],
+		messageProducer:IMessageProducer,
+		participantService: IParticipantService
 	) {
-        this._logger = logger;
+		this._logger = logger;
 		this._oracleFinder = oracleFinder;
 		this._oracleProviders = oracleProviders;
-        this._messageProducer = messageProducer;
-        this._participantService = participantService;
-    }
-
-    async init(): Promise<void> {
-		try {
-            this._oracleFinder.init();
-            this._logger.debug("Oracle finder initialized");
-            for await (const oracle of this._oracleProviders) {
-                await oracle.init();
-                this._logger.debug("Oracle provider initialized with type" + oracle.partyType);
-            }
-    	}
-        catch(error) {
-            {
-			    this._logger.fatal("Unable to intialize account lookup aggregate" + error);
-			    throw error;
-		    }
-	    }
-    }
-
-
-    async publishAccountLookUpEvent(message:IAccountLookUpMessage): Promise<void> {
-        const isMessageValid = this.validateMessage(message);
-        if(isMessageValid) {
-            try{
-                await this.handleMessage(message);
-            }
-            catch(error) {
-                this._logger.error(`${message.value.type} ` + error);
-            }
-        }
-    }
-
-    async destroy(): Promise<void> {
-        try{
-            await this._oracleFinder.destroy();
-            for await (const oracle of this._oracleProviders) {
-                oracle.destroy();
-            }
-        } catch(error) {
-            this._logger.fatal("Unable to destroy account lookup aggregate" + error);
-            throw error;
-        }
+		this._messageProducer = messageProducer;
+		this._participantService = participantService;
 	}
 
-    //Participant.
-    async getParticipant({ sourceFspId, partyType, partyId, partySubType, currency, destinationFspId}: ParticipantQueryReceived):Promise<void>{
-        const sourceFsp = await this._participantService.getParticipantInfo(sourceFspId);
+	async init(): Promise<void> {
+		try {
+			this._oracleFinder.init();
+			this._logger.debug("Oracle finder initialized");
+			for await (const oracle of this._oracleProviders) {
+				await oracle.init();
+				this._logger.debug("Oracle provider initialized with type" + oracle.partyType);
+			}
+		}
+		catch(error) {
+			{
+				this._logger.fatal("Unable to intialize account lookup aggregate" + error);
+				throw error;
+			}
+		}
+	}
+
+
+	async publishAccountLookUpEvent(message:IAccountLookUpMessage): Promise<void> {
+		const isMessageValid = this.validateMessage(message);
+		if(isMessageValid) {
+			try{
+				await this.handleMessage(message);
+			}
+			catch(error) {
+				this._logger.error(`${message.value.type} ` + error);
+			}
+		}
+	}
+
+	async destroy(): Promise<void> {
+		try{
+			await this._oracleFinder.destroy();
+			for await (const oracle of this._oracleProviders) {
+				oracle.destroy();
+			}
+		} catch(error) {
+			this._logger.fatal("Unable to destroy account lookup aggregate" + error);
+			throw error;
+		}
+	}
+
+	//Participant.
+	async getParticipant({ sourceFspId, partyType, partyId, partySubType, currency, destinationFspId}: ParticipantQueryReceived):Promise<void>{
+		const sourceFsp = await this._participantService.getParticipantInfo(sourceFspId);
   
-        this.validateParticipant(sourceFsp);
-        
-        const destinationFsp: IParticipant = await this.getDestinationFsp({ partyId, partyType, partySubType, destinationFspId });
-        
-        await this._messageProducer.send({ 
-            fspId: destinationFsp.id,
-            partyType: partyType,
-            partyId: partyId,
-            partySubType: partySubType,
-            currency: currency
-        });
-        
-    }
+		this.validateParticipant(sourceFsp);
+		
+		const destinationFsp: IParticipant = await this.getDestinationFsp({ partyId, partyType, partySubType, destinationFspId });
+		
+		await this._messageProducer.send({ 
+			fspId: destinationFsp.id,
+			partyType: partyType,
+			partyId: partyId,
+			partySubType: partySubType,
+			currency: currency
+		});
+		
+	}
 
-    //Party.
-    async associateParty({ requesterFspId, partyType, partySubType, partyId }: ParticipantAssociationRequestReceived):Promise<void>{
-        const requesterFsp = await this._participantService.getParticipantInfo(requesterFspId);
+	//Party.
+	async associateParty({ requesterFspId, partyType, partySubType, partyId }: ParticipantAssociationRequestReceived):Promise<void>{
+		const requesterFsp = await this._participantService.getParticipantInfo(requesterFspId);
   
-        this.validateParticipant(requesterFsp);
+		this.validateParticipant(requesterFsp);
 
-        const oracleProvider = await this.getOracleProvider(partyType, partySubType);
+		const oracleProvider = await this.getOracleProvider(partyType, partySubType);
 
-        await oracleProvider.associateParty(partyId).catch(error=>{
-            this._logger.error(`Unable to associate party id: ${partyId} ` + error);
-            throw new UnableToAssociatePartyError(error);
-        });
+		await oracleProvider.associateParty(partyId).catch(error=>{
+			this._logger.error(`Unable to associate party id: ${partyId} ` + error);
+			throw new UnableToAssociatePartyError(error);
+		});
 
-        await this._messageProducer.send(null);
-    }
+		await this._messageProducer.send(null);
+	}
 
-    async disassociateParty({ requesterFspId, partyType, partySubType, partyId }: ParticipantDisassociationRequestReceived):Promise<void>{
-        const requesterFsp = await this._participantService.getParticipantInfo(requesterFspId);
+	async disassociateParty({ requesterFspId, partyType, partySubType, partyId }: ParticipantDisassociationRequestReceived):Promise<void>{
+		const requesterFsp = await this._participantService.getParticipantInfo(requesterFspId);
   
-        this.validateParticipant(requesterFsp);
-        
-        const oracleProvider = await this.getOracleProvider(partyType, partySubType);
+		this.validateParticipant(requesterFsp);
+		
+		const oracleProvider = await this.getOracleProvider(partyType, partySubType);
 
-        await oracleProvider.disassociateParty(partyId).catch(error=>{
-            this._logger.error(`Unable to disassociate party id: ${partyId} ` + error);
-            throw new UnableToDisassociatePartyError(error);
-        });
+		await oracleProvider.disassociateParty(partyId).catch(error=>{
+			this._logger.error(`Unable to disassociate party id: ${partyId} ` + error);
+			throw new UnableToDisassociatePartyError(error);
+		});
 
-        await this._messageProducer.send(null);
-    }
+		await this._messageProducer.send(null);
+	}
 
-    async getPartyRequest({ sourceFspId, partyType, partyId, partySubType, currency, destinationFspId }: PartyQueryReceived):Promise<void>{
+	async getPartyRequest({ sourceFspId, partyType, partyId, partySubType, currency, destinationFspId }: PartyQueryReceived):Promise<void>{
 
-        const sourceFsp = await this._participantService.getParticipantInfo(sourceFspId);
+		const sourceFsp = await this._participantService.getParticipantInfo(sourceFspId);
   
-        this.validateParticipant(sourceFsp);
+		this.validateParticipant(sourceFsp);
 
-        const destinationFsp: IParticipant = await this.getDestinationFsp({ partyId, partyType, partySubType, destinationFspId });
+		const destinationFsp: IParticipant = await this.getDestinationFsp({ partyId, partyType, partySubType, destinationFspId });
 
-        await this._messageProducer.send({ 
-            sourceFspId: sourceFspId,
-            destinationFspId: destinationFsp.id,
-            partyType: partyType,
-            partyId: partyId,
-            partySubType: partySubType,
-            currency: currency
-        });
-        
-    }
-    
-    async getPartyResponse({ sourceFspId, destinationFspId, partyType, partyId, partySubType, currency, partyName, partyDoB }: PartyInfoAvailable):Promise<void>{
-        const sourceFsp = await this._participantService.getParticipantInfo(sourceFspId);
-
-
-        this.validateParticipant(sourceFsp);
-
-        const destinationFsp = await this._participantService.getParticipantInfo(destinationFspId);
-
-        this.validateParticipant(destinationFsp);
-
-        await this._messageProducer.send({ 
-            sourceFspId: sourceFspId,
-            destinationFspId: destinationFspId,
-            partyType: partyType,
-            partyId: partyId,
-            partySubType: partySubType,
-            currency: currency,
-            partyName: partyName,
-            partyDoB: partyDoB
-        });
-    }
+		await this._messageProducer.send({ 
+			sourceFspId: sourceFspId,
+			destinationFspId: destinationFsp.id,
+			partyType: partyType,
+			partyId: partyId,
+			partySubType: partySubType,
+			currency: currency
+		});
+		
+	}
+	
+	async getPartyResponse({ sourceFspId, destinationFspId, partyType, partyId, partySubType, currency, partyName, partyDoB }: PartyInfoAvailable):Promise<void>{
+		const sourceFsp = await this._participantService.getParticipantInfo(sourceFspId);
 
 
-    //Private methods.
-    private validateMessage(message:IAccountLookUpMessage): boolean {
-        if(!message.value){
-            this._logger.error(`AccountLookUpEventHandler: publishAccountLookUpEvent: message as an invalid format or value`);
-            return false;
-        }
+		this.validateParticipant(sourceFsp);
+
+		const destinationFsp = await this._participantService.getParticipantInfo(destinationFspId);
+
+		this.validateParticipant(destinationFsp);
+
+		await this._messageProducer.send({ 
+			sourceFspId: sourceFspId,
+			destinationFspId: destinationFspId,
+			partyType: partyType,
+			partyId: partyId,
+			partySubType: partySubType,
+			currency: currency,
+			partyName: partyName,
+			partyDoB: partyDoB
+		});
+	}
 
 
-        if(! Object.values(AccountLookUpEventsType).some(event => event === message?.value?.type)){
-            this._logger.error(`AccountLookUpEventHandler: publishAccountLookUpEvent: message type ${message.value.type} is not a valid event type`);
-            return false;
-        }
+	//Private methods.
+	private validateMessage(message:IAccountLookUpMessage): boolean {
+		if(!message.value){
+			this._logger.error(`AccountLookUpEventHandler: publishAccountLookUpEvent: message as an invalid format or value`);
+			return false;
+		}
 
-        return true;
-    }
 
-    private async handleMessage(message:IAccountLookUpMessage):Promise<void> {
-        let payload;
-        switch(message.value.type){
-            case AccountLookUpEventsType.GetParty:
-                payload = message.value.payload as PartyQueryReceived;
-                await this.getPartyRequest({ 
-                    sourceFspId: payload.sourceFspId, 
-                    partyType: payload.partyType, 
-                    partyId: payload.partyId, 
-                    partySubType: payload.partySubType, 
-                    currency: payload.currency,
-                    destinationFspId: payload.destinationFspId
-                });
-                break;
-            case AccountLookUpEventsType.GetParticipant:
-                payload = message.value.payload as ParticipantQueryReceived;
-                await this.getParticipant({
-                    sourceFspId: payload.sourceFspId, 
-                    partyType: payload.partyType, 
-                    partyId: payload.partyId, 
-                    partySubType: payload.partySubType, 
-                    destinationFspId: payload.destinationFspId, 
-                    currency: payload.currency
-                });
-                break;
-            case AccountLookUpEventsType.AssociateParty:
-                payload = message.value.payload as ParticipantAssociationRequestReceived;
-                await this.associateParty({
-                    requesterFspId: payload.requesterFspId, 
-                    partyType: payload.partyType, 
-                    partyId: payload.partyId, 
-                    partySubType: payload.partySubType
-                });
-                break;
-            case AccountLookUpEventsType.DisassociateParty:
-                payload = message.value.payload as ParticipantDisassociationRequestReceived;
-                await this.disassociateParty({
-                    requesterFspId: payload.requesterFspId, 
-                    partyType: payload.partyType, 
-                    partyId: payload.partyId, 
-                    partySubType: payload.partySubType
-                });
-                break;
-            default:
-                this._logger.error(`AccountLookUpEventHandler: handleMessage: message type ${message.value.type} is not a valid event type`);
-                break;
-            }
-    }
+		if(! Object.values(AccountLookUpEventsType).some(event => event === message?.value?.type)){
+			this._logger.error(`AccountLookUpEventHandler: publishAccountLookUpEvent: message type ${message.value.type} is not a valid event type`);
+			return false;
+		}
 
-    private async getOracleProvider(partyType:string, partySubType?:string): Promise<IOracleProvider> {
+		return true;
+	}
 
-        const oracleProvider = await this._oracleFinder.getOracleProvider(partyType, partySubType).catch(error=>{
-            this._logger.error(`Unable to get oracle for type: ${partyType} ` + error);
-            throw new UnableToGetOracleError(error);
-        });
+	private async handleMessage(message:IAccountLookUpMessage):Promise<void> {
+		let payload;
+		switch(message.value.type){
+			case AccountLookUpEventsType.GetParty:
+				payload = message.value.payload as PartyQueryReceived;
+				await this.getPartyRequest({ 
+					sourceFspId: payload.sourceFspId, 
+					partyType: payload.partyType, 
+					partyId: payload.partyId, 
+					partySubType: payload.partySubType, 
+					currency: payload.currency,
+					destinationFspId: payload.destinationFspId
+				});
+				break;
+			case AccountLookUpEventsType.GetParticipant:
+				payload = message.value.payload as ParticipantQueryReceived;
+				await this.getParticipant({
+					sourceFspId: payload.sourceFspId, 
+					partyType: payload.partyType, 
+					partyId: payload.partyId, 
+					partySubType: payload.partySubType, 
+					destinationFspId: payload.destinationFspId, 
+					currency: payload.currency
+				});
+				break;
+			case AccountLookUpEventsType.AssociateParty:
+				payload = message.value.payload as ParticipantAssociationRequestReceived;
+				await this.associateParty({
+					requesterFspId: payload.requesterFspId, 
+					partyType: payload.partyType, 
+					partyId: payload.partyId, 
+					partySubType: payload.partySubType
+				});
+				break;
+			case AccountLookUpEventsType.DisassociateParty:
+				payload = message.value.payload as ParticipantDisassociationRequestReceived;
+				await this.disassociateParty({
+					requesterFspId: payload.requesterFspId, 
+					partyType: payload.partyType, 
+					partyId: payload.partyId, 
+					partySubType: payload.partySubType
+				});
+				break;
+			default:
+				this._logger.error(`AccountLookUpEventHandler: handleMessage: message type ${message.value.type} is not a valid event type`);
+				break;
+			}
+	}
 
-        if(!oracleProvider) {
-            this._logger.debug(`No oracle provider found for partyType: ${partyType}`);
-            throw new UnableToGetOracleProviderError(`Oracle provider not found for partyType: ${partyType}`);
-        }
+	private async getOracleProvider(partyType:string, partySubType?:string): Promise<IOracleProvider> {
+
+		const oracleProvider = await this._oracleFinder.getOracleProvider(partyType, partySubType).catch(error=>{
+			this._logger.error(`Unable to get oracle for type: ${partyType} ` + error);
+			throw new UnableToGetOracleError(error);
+		});
+
+		if(!oracleProvider) {
+			this._logger.debug(`No oracle provider found for partyType: ${partyType}`);
+			throw new UnableToGetOracleProviderError(`Oracle provider not found for partyType: ${partyType}`);
+		}
 
 		return oracleProvider;
 	}
-    
-    private validateParticipant(participant: IParticipant | null):void{
+	
+	private validateParticipant(participant: IParticipant | null):void{
 
-        if(!participant) {
-            throw new NoSuchParticipantFspIdError(`fspId does not exist`);
-        }
+		if(!participant) {
+			throw new NoSuchParticipantFspIdError(`fspId does not exist`);
+		}
 
-        if(!participant.isActive) {
-            throw new RequiredParticipantIsNotActive(`fspId:${participant.id} is not active`);
-        }
-        
-    }
+		if(!participant.isActive) {
+			throw new RequiredParticipantIsNotActive(`fspId:${participant.id} is not active`);
+		}
+		
+	}
 
-    private async getDestinationFsp({ partyId, partyType, partySubType, destinationFspId }: { partyId:string, partyType:string, partySubType?:string, destinationFspId?: string }):Promise<IParticipant>{
-        
-        let destinationFsp: IParticipant|null = null;
+	private async getDestinationFsp({ partyId, partyType, partySubType, destinationFspId }: { partyId:string, partyType:string, partySubType?:string, destinationFspId?: string }):Promise<IParticipant>{
+		
+		let destinationFsp: IParticipant|null = null;
 
-        // In this case, the sourceFspId already knows the destinationFspId,
-        // so we just need to validate it
-        if(destinationFspId) {
-            const participant = await this._participantService.getParticipantInfo(destinationFspId);
-            
-            this.validateParticipant(participant);
+		// In this case, the sourceFspId already knows the destinationFspId,
+		// so we just need to validate it
+		if(destinationFspId) {
+			const participant = await this._participantService.getParticipantInfo(destinationFspId);
+			
+			this.validateParticipant(participant);
 
-            if(participant) {
-                destinationFsp = participant; 
-            }
-        } else {
-     
-            
-            destinationFsp = await this.getParticipantFromOracle(partyId, partyType, partySubType);
-        }  
+			if(participant) {
+				destinationFsp = participant; 
+			}
+		} else {
+			destinationFsp = await this.getParticipantFromOracle(partyId, partyType, partySubType);
+		}  
 
-        if(!destinationFsp){
-            throw new NoSuchParticipantFspIdError(`partyId:${partyId} has no existing fspId owner`);
-        }
+		if(!destinationFsp){
+			throw new NoSuchParticipantFspIdError(`partyId:${partyId} has no existing fspId owner`);
+		}
 
-        return destinationFsp;
-    }
+		return destinationFsp;
+	}
 
-    private async getParticipantFromOracle(partyId:string, partyType:string, partySubType?:string): Promise<IParticipant|null> {
+	private async getParticipantFromOracle(partyId:string, partyType:string, partySubType?:string): Promise<IParticipant|null> {
 
-        const oracle = await this._oracleFinder.getOracleProvider(partyType, partySubType);
-        
-        if(!oracle) {
-            throw new NoSuchOracleProviderError(`oracle for ${partyType} not found`);
-        }
+		const oracle = await this._oracleFinder.getOracleProvider(partyType, partySubType);
+		
+		if(!oracle) {
+			throw new NoSuchOracleProviderError(`oracle for ${partyType} not found`);
+		}
 
-        // We can have more than one FSP that owns this partyId
-        const fspId = await oracle.getParticipant(partyId);
+		// We can have more than one FSP that owns this partyId
+		const fspId = await oracle.getParticipant(partyId);
 
-        if(!(fspId)) {
-            throw new NoSuchParticipantFspIdError(`partyId:${partyId} has no existing fspId owner`);
-        }
+		if(!(fspId)) {
+			throw new NoSuchParticipantFspIdError(`partyId:${partyId} has no existing fspId owner`);
+		}
 
-        // The participants BC will return a participant's info  
-        // more info than just a regular ID, example:
-        // {
-        //     id: string;
-        //     isActive: Boolean;
-        // }
-        const participant = await this._participantService.getParticipantInfo(fspId);
-        
+		// The participants BC will return a participant's info  
+		// more info than just a regular ID, example:
+		// {
+		//     id: string;
+		//     isActive: Boolean;
+		// }
+		const participant = await this._participantService.getParticipantInfo(fspId);
+		
 
-        this.validateParticipant(participant);
+		this.validateParticipant(participant);
 
-        return participant;
-    }
+		return participant;
+	}
 }
